requirements.txt
accelerate==0.16.0
albumentations==1.3.0
imutils==0.5.4
loguru==0.6.0
matplotlib==3.7.1
notifiers==1.3.3
numpy==1.23.5
omegaconf==2.3.0
opencv_python_headless==4.7.0.68
optuna==3.1.0
pandas==1.5.3
Pillow==9.4.0
pymongo==4.3.3
pytorch_metric_learning==2.0.0
PyTurboJPEG==1.7.0
PyYAML==6.0
scikit_learn==1.2.2
timm==0.6.12
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.1
typer==0.7.0
